LanguageSelector refactored into own component
