+++ src/teacher/LabeledInputFields.sc.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { darkBlue } from "../components/colors";
 import "@reach/listbox/styles.css";
 
-export const CohortFormInputFields = styled.div`
+export const LabeledInputFields = styled.div`
   .input-container {
     margin-top: 1em;
   }
